[nan-1328] reuse methods
