+++ integration-templates/xero/actions/xero-create-contact.ts
@@ -1,5 +1,6 @@
 import type { NangoAction, Contact, ContactActionResponse, FailedContact, ActionErrorResponse } from '../../models';
 import { getTenantId } from '../helpers/get-tenant-id.js';
+import { toXeroContact, toContact } from '../mappers/to-contact.js';
 
 export default async function runAction(nango: NangoAction, input?: Contact[]): Promise<ContactActionResponse> {
     const tenant_id = await getTenantId(nango);
@@ -28,7 +29,7 @@ export default async function runAction(nango: NangoAction, input?: Contact[]): 
             summarizeErrors: 'false'
         },
         data: {
-            Contacts: input.map(mapContactToXero)
+            Contacts: input.map(toXeroContact)
         }
     };
 
@@ -49,103 +50,16 @@ export default async function runAction(nango: NangoAction, input?: Contact[]): 
     const succeededContacts = contacts.filter((x: any) => !x.HasValidationErrors);
 
     const response = {
-        succeededContacts: succeededContacts.map(mapXeroContact),
+        succeededContacts: succeededContacts.map(toContact),
         failedContacts: failedContacts.map(mapFailedXeroContact)
     } as ContactActionResponse;
 
     return response;
 }
 
-// Maps a Nango Contact to a Xero Contact
-function mapContactToXero(contact: Contact) {
-    const xeroContact: any = {};
-
-    if (contact.name) {
-        xeroContact['Name'] = contact.name;
-    }
-
-    if (contact.external_id) {
-        xeroContact['ContactNumber'] = contact.external_id;
-    }
-
-    if (contact.tax_number) {
-        xeroContact['TaxNumber'] = contact.tax_number;
-    }
-
-    if (contact.email) {
-        xeroContact['EmailAddress'] = contact.email;
-    }
-
-    if (contact.phone) {
-        xeroContact['Phones'] = [
-            {
-                PhoneType: 'DEFAULT',
-                PhoneNumber: contact.phone
-            }
-        ];
-    }
-
-    if (contact.address_line_1 || contact.address_line_2 || contact.city || contact.zip || contact.country || contact.state) {
-        const address: any = { AddressType: 'POBOX' };
-        if (contact.address_line_1) {
-            address['AddressLine1'] = contact.address_line_1;
-        }
-        if (contact.address_line_2) {
-            address['AddressLine2'] = contact.address_line_2;
-        }
-        if (contact.city) {
-            address['City'] = contact.city;
-        }
-        if (contact.zip) {
-            address['PostalCode'] = contact.zip;
-        }
-        if (contact.country) {
-            address['Country'] = contact.country;
-        }
-        if (contact.state) {
-            address['Region'] = contact.state;
-        }
-
-        xeroContact['Addresses'] = [address];
-    }
-
-    return xeroContact;
-}
-
 function mapFailedXeroContact(xeroContact: any): FailedContact {
-    const failedContact = mapXeroContact(xeroContact) as FailedContact;
+    const failedContact = toContact(xeroContact) as FailedContact;
     failedContact.validation_errors = xeroContact.ValidationErrors;
 
     return failedContact;
-}
-
-function mapXeroContact(xeroContact: any): Contact {
-    // Find Street address & default phone object, if they exist
-    let streetAddress = xeroContact.Addresses.filter((x: any) => x.AddressType === 'POBOX')[0];
-    let defaultPhone = xeroContact.Phones.filter((x: any) => x.PhoneType === 'DEFAULT')[0];
-
-    streetAddress = streetAddress ? streetAddress : {};
-    defaultPhone = defaultPhone ? defaultPhone : {};
-
-    let formattedPhoneNumber: any = null;
-    if (defaultPhone.PhoneNumber) {
-        formattedPhoneNumber = defaultPhone.PhoneCountryCode
-            ? `+${defaultPhone.PhoneCountryCode}${defaultPhone.PhoneAreaCode}${defaultPhone.PhoneNumber}`
-            : `${defaultPhone.PhoneAreaCode}${defaultPhone.PhoneNumber}`;
-    }
-
-    return {
-        id: xeroContact.ContactID,
-        name: xeroContact.Name,
-        external_id: xeroContact.ContactNumber,
-        tax_number: xeroContact.TaxNumber,
-        email: xeroContact.EmailAddress,
-        address_line_1: streetAddress.AddressLine1,
-        address_line_2: streetAddress.AddressLine2,
-        city: streetAddress.City,
-        zip: streetAddress.PostalCode,
-        country: streetAddress.Country,
-        state: streetAddress.Region,
-        phone: formattedPhoneNumber
-    } as Contact;
 }
